+++ src/creature.rs
@@ -9,7 +9,7 @@ use crate::error::{WeaselError, WeaselResult};
 use crate::event::{Event, EventKind, EventProcessor, EventQueue, EventTrigger};
 use crate::metric::system::*;
 use crate::round::RoundState;
-use crate::space::Position;
+use crate::space::{Position, PositionClaim};
 use crate::team::{EntityAddition, TeamId, TeamRules};
 use crate::util::Id;
 #[cfg(feature = "serialization")]
@@ -284,7 +284,10 @@ impl<R: BattleRules + 'static> Event<R> for CreateCreature<R> {
             return Err(WeaselError::DuplicatedCreature(self.id.clone()));
         }
         // Check position.
-        if !battle.space().check_move(None, &self.position) {
+        if !battle.space().check_move(
+            PositionClaim::Spawn(&EntityId::Creature(self.id.clone())),
+            &self.position,
+        ) {
             return Err(WeaselError::PositionError(None, self.position.clone()));
         }
         Ok(())
@@ -314,10 +317,11 @@ impl<R: BattleRules + 'static> Event<R> for CreateCreature<R> {
             abilities,
         };
         // Take the position.
-        battle
-            .state
-            .space
-            .move_entity(None, &self.position, &mut battle.metrics.write_handle());
+        battle.state.space.move_entity(
+            PositionClaim::Spawn(&EntityId::Creature(self.id.clone())),
+            Some(&self.position),
+            &mut battle.metrics.write_handle(),
+        );
         // Notify the rounds module.
         battle.state.rounds.on_actor_added(
             &creature,
@@ -558,6 +562,9 @@ where
 }
 
 /// Event to remove a creature from the battle.
+///
+/// If the creature is the current actor, its round will be terminated.\
+/// The creature will be removed from the corresponding team and its position will be freed.
 #[cfg_attr(feature = "serialization", derive(Serialize, Deserialize))]
 pub struct RemoveCreature<R: BattleRules> {
     #[cfg_attr(
@@ -636,11 +643,17 @@ impl<R: BattleRules + 'static> Event<R> for RemoveCreature<R> {
             }
         }
         // Remove the creature.
-        battle
+        let creature = battle
             .state
             .entities
             .remove_creature(&self.id)
             .unwrap_or_else(|err| panic!("constraint violated: {:?}", err));
+        // Free the position.
+        battle.state.space.move_entity(
+            PositionClaim::Movement(&creature as &dyn Entity<R>),
+            None,
+            &mut battle.metrics.write_handle(),
+        );
     }
 
     fn kind(&self) -> EventKind {