+++ src/space.rs
@@ -1,11 +1,11 @@
 //! Module for the spatial dimension.
 
-use crate::battle::Battle;
-use crate::battle::BattleRules;
-use crate::entity::{Entity, EntityId};
+use crate::battle::{Battle, BattleRules};
+use crate::entity::{Entities, Entity, EntityId};
 use crate::error::{WeaselError, WeaselResult};
 use crate::event::{Event, EventKind, EventProcessor, EventQueue, EventTrigger};
 use crate::metric::WriteMetrics;
+use crate::round::Rounds;
 #[cfg(feature = "serialization")]
 use serde::{Deserialize, Serialize};
 use std::any::Any;
@@ -28,23 +28,23 @@ impl<R: BattleRules> Space<R> {
     }
 
     /// See [check_move](SpaceRules::check_move).
-    pub(crate) fn check_move(
+    pub(crate) fn check_move<'a>(
         &self,
-        entity: Option<&dyn Entity<R>>,
+        claim: PositionClaim<'a, R>,
         position: &Position<R>,
     ) -> bool {
-        self.rules.check_move(&self.model, entity, position)
+        self.rules.check_move(&self.model, claim, position)
     }
 
     /// See [move_entity](SpaceRules::move_entity).
-    pub(crate) fn move_entity(
+    pub(crate) fn move_entity<'a>(
         &mut self,
-        entity: Option<&dyn Entity<R>>,
-        position: &Position<R>,
+        claim: PositionClaim<'a, R>,
+        position: Option<&Position<R>>,
         metrics: &mut WriteMetrics<R>,
     ) {
         self.rules
-            .move_entity(&mut self.model, entity, position, metrics);
+            .move_entity(&mut self.model, claim, position, metrics);
     }
 
     /// Returns the space model.
@@ -78,6 +78,13 @@ pub trait SpaceRules<R: BattleRules> {
     /// See [SpaceSeed](type.SpaceSeed.html).
     type SpaceSeed: Clone + Debug + Serialize + for<'a> Deserialize<'a>;
 
+    #[cfg(not(feature = "serialization"))]
+    /// See [SpaceAlteration](type.SpaceAlteration.html).
+    type SpaceAlteration: Clone + Debug;
+    #[cfg(feature = "serialization")]
+    /// See [SpaceAlteration](type.SpaceAlteration.html).
+    type SpaceAlteration: Clone + Debug + Serialize + for<'a> Deserialize<'a>;
+
     /// See [SpaceModel](type.SpaceModel.html).
     type SpaceModel;
 
@@ -86,14 +93,13 @@ pub trait SpaceRules<R: BattleRules> {
 
     /// Checks if a given entity can move into a new position.
     ///
-    /// In the case entity is empty, it is assumed that `position` will be the starting one of a
-    /// new entity.
+    /// The claim tells in which context the entity is trying to acquire the position.
     ///
     /// The provided implementation accepts every move.
-    fn check_move(
+    fn check_move<'a>(
         &self,
         _model: &Self::SpaceModel,
-        _entity: Option<&dyn Entity<R>>,
+        _claim: PositionClaim<'a, R>,
         _position: &Self::Position,
     ) -> bool {
         true
@@ -101,21 +107,22 @@ pub trait SpaceRules<R: BattleRules> {
 
     /// Moves an entity into a new position.
     ///
-    /// Position's correctness will be validated beforehand with `check_move`.\
-    /// In the case entity is empty, it is assumed that `position` will be the starting one of a
-    /// new entity.
+    /// Position's correctness will be validated beforehand with `check_move`,
+    /// unless it is `None`.
+    /// An empty position means that the entity is disappearing from the battle.\
+    /// The claim tells in which context the entity is trying to acquire the position.\
     ///
     /// The provided implementation does nothing.
-    fn move_entity(
+    fn move_entity<'a>(
         &self,
         _model: &mut Self::SpaceModel,
-        _entity: Option<&dyn Entity<R>>,
-        _position: &Self::Position,
+        _claim: PositionClaim<'a, R>,
+        _position: Option<&Self::Position>,
         _metrics: &mut WriteMetrics<R>,
     ) {
     }
 
-    /// Translate an entity from one space model to another one.
+    /// Translates an entity from one space model to another one.
     ///
     /// This method must apply the necessary changes to the entity's position and to the new model
     /// so that positions consistency is preserved.
@@ -126,6 +133,22 @@ pub trait SpaceRules<R: BattleRules> {
         _model: &Self::SpaceModel,
         _new_model: &mut Self::SpaceModel,
         _entity: &mut dyn Entity<R>,
+        _event_queue: &mut Option<EventQueue<R>>,
+        _metrics: &mut WriteMetrics<R>,
+    ) {
+    }
+
+    /// Changes the current space model, starting from the information contained in `alteration`.
+    ///
+    /// Consequences of this change should be applied by registering events inside `event_queue`.
+    ///
+    /// The provided implementation does nothing.
+    fn alter_space(
+        &self,
+        _entities: &Entities<R>,
+        _rounds: &Rounds<R>,
+        _model: &mut Self::SpaceModel,
+        _alteration: &Self::SpaceAlteration,
         _event_queue: &mut Option<EventQueue<R>>,
         _metrics: &mut WriteMetrics<R>,
     ) {
@@ -150,6 +173,21 @@ pub type SpaceSeed<R> = <<R as BattleRules>::SR as SpaceRules<R>>::SpaceSeed;
 /// space and all environment hazards.\
 /// An example of space model is a matrix containing the position of all pieces in a game of chess.
 pub type SpaceModel<R> = <<R as BattleRules>::SR as SpaceRules<R>>::SpaceModel;
+
+/// Represents an alteration to the space model.
+///
+/// It is used to evolve the battle's current space model. This alteration should only contain
+/// the direct changes to the model. All side effects deriving from such change have to be
+/// implemented in the space rules `alter_space` method.
+pub type SpaceAlteration<R> = <<R as BattleRules>::SR as SpaceRules<R>>::SpaceAlteration;
+
+/// Represents an entity's claim to a given position.
+pub enum PositionClaim<'a, R: BattleRules> {
+    /// The entity is spawning.
+    Spawn(&'a EntityId<R>),
+    /// The entity wants to change its position.
+    Movement(&'a dyn Entity<R>),
+}
 
 /// An event to move an entity from its position to a new one.
 #[cfg_attr(feature = "serialization", derive(Serialize, Deserialize))]
@@ -225,7 +263,10 @@ impl<R: BattleRules + 'static> Event<R> for MoveEntity<R> {
             .entity(&self.id)
             .ok_or_else(|| WeaselError::EntityNotFound(self.id.clone()))?;
         // Check position.
-        if !battle.space().check_move(Some(entity), &self.position) {
+        if !battle
+            .space()
+            .check_move(PositionClaim::Movement(entity), &self.position)
+        {
             return Err(WeaselError::PositionError(
                 Some(entity.position().clone()),
                 self.position.clone(),
@@ -243,8 +284,8 @@ impl<R: BattleRules + 'static> Event<R> for MoveEntity<R> {
             .unwrap_or_else(|| panic!("constraint violated: entity {:?} not found", self.id));
         // Take the new position.
         battle.state.space.move_entity(
-            Some(entity),
-            &self.position,
+            PositionClaim::Movement(entity),
+            Some(&self.position),
             &mut battle.metrics.write_handle(),
         );
         // Update the entity.
@@ -406,6 +447,112 @@ where
     fn event(&self) -> Box<dyn Event<R>> {
         Box::new(ResetSpace {
             seed: self.seed.clone(),
+        })
+    }
+}
+
+/// Event to alter the space model.
+///
+/// Alterations to the space model might have consequences on entities' or on other
+/// aspects of the battle.
+#[cfg_attr(feature = "serialization", derive(Serialize, Deserialize))]
+pub struct AlterSpace<R: BattleRules> {
+    #[cfg_attr(
+        feature = "serialization",
+        serde(bound(
+            serialize = " SpaceAlteration<R>: Serialize",
+            deserialize = " SpaceAlteration<R>: Deserialize<'de>"
+        ))
+    )]
+    alteration: SpaceAlteration<R>,
+}
+
+impl<R: BattleRules> AlterSpace<R> {
+    /// Returns a trigger for this event.
+    pub fn trigger<P: EventProcessor<R>>(
+        processor: &mut P,
+        alteration: SpaceAlteration<R>,
+    ) -> AlterSpaceTrigger<R, P> {
+        AlterSpaceTrigger {
+            processor,
+            alteration,
+        }
+    }
+
+    /// Returns the alteration to be applied to the space model.
+    pub fn alteration(&self) -> &SpaceAlteration<R> {
+        &self.alteration
+    }
+}
+
+impl<R: BattleRules> Debug for AlterSpace<R> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+        write!(f, "AlterSpace {{ alteration: {:?} }}", self.alteration)
+    }
+}
+
+impl<R: BattleRules> Clone for AlterSpace<R> {
+    fn clone(&self) -> Self {
+        AlterSpace {
+            alteration: self.alteration.clone(),
+        }
+    }
+}
+
+impl<R: BattleRules + 'static> Event<R> for AlterSpace<R> {
+    fn verify(&self, _battle: &Battle<R>) -> WeaselResult<(), R> {
+        Ok(())
+    }
+
+    fn apply(&self, battle: &mut Battle<R>, event_queue: &mut Option<EventQueue<R>>) {
+        let rules = &battle.state.space.rules;
+        // Apply the alteration.
+        rules.alter_space(
+            &battle.state.entities,
+            &battle.state.rounds,
+            &mut battle.state.space.model,
+            &self.alteration,
+            event_queue,
+            &mut battle.metrics.write_handle(),
+        );
+    }
+
+    fn kind(&self) -> EventKind {
+        EventKind::AlterSpace
+    }
+
+    fn box_clone(&self) -> Box<dyn Event<R>> {
+        Box::new(self.clone())
+    }
+
+    fn as_any(&self) -> &dyn Any {
+        self
+    }
+}
+
+/// Trigger to build and fire a `AlterSpace` event.
+pub struct AlterSpaceTrigger<'a, R, P>
+where
+    R: BattleRules,
+    P: EventProcessor<R>,
+{
+    processor: &'a mut P,
+    alteration: SpaceAlteration<R>,
+}
+
+impl<'a, R, P> EventTrigger<'a, R, P> for AlterSpaceTrigger<'a, R, P>
+where
+    R: BattleRules + 'static,
+    P: EventProcessor<R>,
+{
+    fn processor(&'a mut self) -> &'a mut P {
+        self.processor
+    }
+
+    /// Returns a `AlterSpace` event.
+    fn event(&self) -> Box<dyn Event<R>> {
+        Box::new(AlterSpace {
+            alteration: self.alteration.clone(),
         })
     }
 }